CI.Linux.install: create /dev/snd if not existent only

+++ src/appveyor/0_install_linux4packaging.sh
@@ -20,7 +20,7 @@ sudo chmod +x /usr/local/bin/appimage-builder
 # give the current user access to docker socket (requires log out+in to activate)
 sudo adduser $(id -un) docker
 # fix docker error about missing snd device:
-sudo ln /dev/null /dev/snd
+[ -f /dev/snd ] || sudo ln /dev/null /dev/snd
 # pull docker img now, prevents errors later
 #docker pull appimagecrafters/tests-env:ubuntu-bionic
 #docker pull appimagecrafters/tests-env:centos-7
